LIBITD-568. Make availability selection a big grid of check_boxes
This makes the availability_times selection step UI a table of days
and hours which users can select from. The db schema is changed to keep
available times in a single table, with day and start hour recorded ( we are
assuming hour long chunks  )

This also removed the nested_form gem and adds the simple_form and cocoon gems (
which nested_form was using under the hood ). This hopefully simplifies the form
helper fun.

https://issues.umd.edu/browse/LIBITD-568