LIBITD-667 Add position_title to work_experience
There was a request to add a Name of Employer to the work experience. However, there
was a "Name" field, but I think this got confused with a position_title. I've added a position
title field, updated the labels in the form, and added things to the test seed data.

https://issues.umd.edu/browse/LIBITD-667