LIBITD-878 Add per_page option to the prospect index page
This adds a select box that allows users to change the number of records
displayed on the index page. It is inside the pagination section.
This was added to make it easier for admins to filter and bulk delete.
Since delete changes the number of records, a 'refresh' after delete was also
added.
There's also a bunch of smalle changes the rubocop added

http://issues.umd.edu/browse/LIBITD-878