LIBITD-1135. Updated Paperclip gem to v5.3.0
Updated Paperclip gem from v5.0.0 to v5.3.0 to address gem
vulnerabilities.

Ran "bundle update paperclip" to update related gems.

Fixed a test that began failing due to better file content detection
in the new Rails version (see
https://github.com/thoughtbot/paperclip/pull/2270). Prior to the
upgrade, the test was actually providing a PDF file, but the file was
being rejected because the provided MIME type was given as "text/html".
After the upgrade, the provided MIME type is not trusted, and the file
itself is examined. Since the file provided in the test is actually a
PDF file, the file is correctly identified as a PDF, and is accepted, so
the test failed. Updated the test to use a non-PDF file.

Also removed "before_validation" code in the
"student-applications/app/models/resume.rb" file, as the default
content type detector in Paperclip now performs a similar check
(see https://github.com/thoughtbot/paperclip/blob/v5.3.0/lib/paperclip/content_type_detector.rb).

https://issues.umd.edu/browse/LIBITD-1135