LIBITD-1074. Tests now use Selenium and headless Chrome
Modified the test infrastructure and tests to use Selenium and
headless Chrome instead of poltergeist (and phantomjs).

Added "selenium-webdriver" gem to enable Capybara to use Selenium.

Added "chromedriver-helper" gem to download and use chromedriver if
it is not present on the system.

Added "lib/animations.rb" as Rack middleware to turn off CSS
animations/transitions which were causing timing/visibility issues
when running in Selenium.

Updated tests relying on the Poltergeist API to use the Webdriver API
and other quirks of the Poltergeist/phantomjs setup, including the
following:

* Updated "drag_until" method in test/test_helper.rb
* Replaced "trigger(:click)" with just "click"
* Added "no_animations.rb" Rack middleware turn off CSS animations when
  testing
* Set absolute path when referencing files
* Fixed text verification in comment_confirmation_test.rb

https://issues.umd.edu/browse/LIBITD-1074