LIBITD-1219. Modified filter to use AND between selection criteria
Modified filtering so that different criteria types (Class Status,
Semester, Preferred Libraries, etc.) will be combined using "AND".
The resulting records must have a matching entry for each criteria type.

If multiple values within a criteria types are selected, the entries
for that type are combined using "OR", i.e., the resulting records must
have at least one of the selected entries.

The behavior of the "Available Times" criteria has not been changed --
multiple selections are "ANDed" together.

Fixed a minor typo in the app/views/prospects/_filter_modal.html.erb
that I found when working on test cases.

Added a few more prospects to the test fixtures, and renamed some of the
fixtures to be more consistent.

Removed "prospects" entries from "enumerations.yml" fixture, because
they do not appear to be necessary.

https://issues.umd.edu/browse/LIBITD-1219