LIBITD-1220. Updated files based on Rails 5.0.7 diff
Used http://railsdiff.org/ to compare Rails 4.2.10 files against
Rails 5.0.7 files, and adjusted based on changes.

When adjusting, used a freshly built Rail 5.0.7 to compare against,
using http://railsdiff.org/ to determine what files to examine.

Added tmp/.keep file because it was mentioned in the .gitignore file,
but was not present.

https://issues.umd.edu/browse/LIBITD-1220