LIBITD-1231. Modified strong parameter handling for arrays
The "params" provided by Rails may contain hash values which use
numeric keys. For example, when passing in the addresses, params are
indexed numerically (0, 1, etc.).

This is a problem for the Rails strong parameter handling, as it does
not support arbitrary keys. Additionally, in Rails 4.2.10, the
ActionController:Parameters object could be treated as a hash, but
this is no longer allowed in Rails 5.0.7.

The whitelisted parameters are specified by the "whitelisted_attrs"
method, returns a hash that has the allowed parameters for array
attributes listed as "addresses_attributes", "phone_numbers_attributes",
etc.

The "convert_attributes_param_to_safe_hash" generates a hash containing
only the whitelisted parameters for an ActionController:Parameters
object. This allows a hash to be retrieved from the object, while
only allowing the whitelisted parameters through.

https://issues.umd.edu/browse/LIBITD-1231