LIBITD-1231. Modified "day_times" updates in test
In Rails 5.0.7, the "update_attribute" only actually saves if it detects
a change, which it doesn't seem to do for associations such as
"day_times". This was apparently fixed in Rails 5.2.0
(see https://github.com/rails/rails/pull/27780).

Added a "set_day_times" method that sets the attribute using the setter
and then calls the "save!" method.

https://issues.umd.edu/browse/LIBITD-1231