LIBITD-1236. Removed failing validation test.
In Rails 5.0, "belongs_to" triggers a validation error by default if the
association is not present. The AvailableTime model has a
"belongs_to: prospect". Since this did not seem to be optional, did not
fix by simply making it optional.

Instead, the test itself seems to be a holdover when the AvailableTime
model did not have a "belongs_to: prospect". So removing the test seemed
like the best option.

https://issues.umd.edu/browse/LIBITD-1236