LIBITD-1236. Replaced symbols with strings in hash retrieval
The "session[:cas]" hash has string keys, so modified the code to
use strings instead of symbols. The symbols apparently worked previously
because of a "with_indifferent_access" conversion in
app/controllers/application_controller.rb, but this conversion does not
seem to be available when the application reaches the "delayed_job_web"
code.

https://issues.umd.edu/browse/LIBITD-1236