LIBITD-1241. Modified .rubocop.yml to include tests and fixed offenses
The test/ directory was being completely excluded, which is not
consistent with other projects. Modified .rubocop.yml to include tests,
and fixed resulting Rubocop offenses.

Also renamed test/models/resume.rb to test/models/resume_test.rb for
consistency, and so that it would get tested. Changed reference to
"user" in test with "resume", as that seemed correct, and the test
fails otherwise.

https://issues.umd.edu/browse/LIBITD-1241