LIBITD-1441. Replaced "chromedriver-helper" with "webdrivers"
The "chromedriver-helper" gem is now obsolete, and does not work with
the latest versions of Chrome. Replaced with "webdrivers" which is
the recommended replacement.

Locking the "webdrivers" version to v3.9, as v4 requires Ruby 2.3,
which we have not yet reached.

This change is needed as of August 30, 2019, in order for the
Selenium-based tests to pass.

https://issues.umd.edu/browse/LIBITD-1441