LIBITD-2048. Modified soft-delete to change "updated_at" field.
The "deactivate" functionality in the "ProspectsController" uses
"update_all" to "soft-delete" the prospects by setting the "suppressed"
flag to "true". The "update_all" method does _not_, of itself, change
the "updated_at" field, so modified it to do so.

This ensures that we can implement a grace period for purging.

https://issues.umd.edu/browse/LIBITD-2048