LIBITD-2113. Dropped paperclip and "reset" database schema.rb (Stage 2)
Removed the "paperclip" gem and related configuration files. This
causes issues with the database migration files, because there was
a migration that depended on paperclip functionality that is no longer
present.

Following the guidance in

https://naturaily.com/blog/remove-old-migrations-ruby-on-rails

created a new migration consisting of the contents of the "schema.rb"
file, and set the version timestamp to the last migration. Then
deleted all prior migrations.

https://issues.umd.edu/browse/LIBITD-2113