LIBITD-2593. Zeitwork autoload should ignore "lib/no_animations.rb"
The Jenkins build was failing with the following error:

```
Coverage report Rcov style generated for Minitest to /apps/ci/jenkins/jenkins-env/workspace/applications_feature_LIBITD-2593/coverage/rcov
Stopped processing SimpleCov as a previous error not related to SimpleCov has been detected
/usr/local/bundle/gems/zeitwerk-2.7.3/lib/zeitwerk/cref.rb:62:in `const_get': uninitialized constant NoAnimations (NameError)

    @mod.const_get(@cname, false)
        ^^^^^^^^^^
	from /usr/local/bundle/gems/zeitwerk-2.7.3/lib/zeitwerk/cref.rb:62:in `get'
```

This is related to use of the “Zeitwerk” loader in Rails 7.1, see
<https://guides.rubyonrails.org/v7.1/classic_to_zeitwerk_howto.html>.

The following command shows that the “NoAnimations” constant
(in “lib/no_animations.rb”) can’t be found:

```
$ rails zeitwerk:check
...
Hold on, I am eager loading the application.
expected file lib/no_animations.rb to define constant NoAnimations, but didn't
``

Zeitwerk apparently expects the “lib/no_animations.rb” to define a
constant named “NoAnimations” (which is not the purpose of the file).

Tried removed the “lib/no_animations.rb” file (and the reference to it
in “config/environments/test.rb”), but this seemed to cause the system
tests to intermittently fail more often, so it seemed better to leave it
in.

Fixed by adding the “lib/no_animations.rb” file to an autoload “ignore”
list in “config/application.rb”.

https://umd-dit.atlassian.net/browse/LIBITD-2593