LIBITD-2593. Implement Post/Redirect/Get (PRG) pattern for prospects
Modified the “create” method in "prospects_controller.rb" to do a
GET redirect after POST, when the prospect has not yet been saved.

This ensures that doing a browser refresh on the page keeps the
user on the same page (this emulates the Rails 6.x Turbolinks behavior).

Updated a system test, because the "current_step" is now getting
properly updated (also verified the expected behavior of the test
manually).

https://umd-dit.atlassian.net/browse/LIBITD-2593