LIBITD-2593. Customize Minitest reporter for Jenkins
In Rails 7.2.x, the Rails unit and system tests run in separate
processes.

The JUnitReporter used by Jenkins, by default empties the "test/reports"
directory before running the tests. This means that only the results
for either the unit or system tests are preserved (depending on which
one ran last).

Modified the "test_helper.rb" configuration to use a JUniteReporter
configuration that does not empty the test reports directories, whenever
the "CI" environment variable is set.

Modified the "Jenkinsfile" to set the "CI" environment variable
(replacing the "MINITEST_REPORTER" environment variable), and also
updated the email to report the number of failed tests.

iUsed “CI” for the environment variable, because that is also
used in the stock Rails “config/environments/test.rb” to denote
that the tests are running on a continuous integration system.

Add "test/reports" to ".gitignore" to prevent any tests reports from
being added to Git.

https://umd-dit.atlassian.net/browse/LIBITD-2593