LIBITD-2593. Fix "Setting priority countries" deprecation
Fixed "Setting priority countries with the 1.x syntax" deprecations when
running the tests by updating the "simple_form" and "country_select"
gem versions.

Note that the "priority" setting in
"app/views/prospects/_address_fields.html.erb" does not need to be
changed, because in the "simple_form" gem, "priority" is still used.

https://umd-dit.atlassian.net/browse/LIBITD-2593