LIBITD-2593. Added additional information on OpenSSL workaround
After evaluating this is due to a mismatch between OpenSSL 3.6.0, and
Ruby 3.4. Possible workarounds are discussed in Issue 55886 in
https://github.com/rails/rails

The Dockerfile.ci (and Dockerfile) will be based on a Ruby 3.4 Docker
image that uses OpenSSL 3.5.4, and adding the "openssl" to the Gemfile
as suggested causes a build error. So keeping this change in the
"development.rb" file as the simplest solution for the moment.

https://umd-dit.atlassian.net/browse/LIBITD-2593