LIBITD-2593. Tweaked steps in "docs/TestPlan.md"
Modified the "docs/TestPlan.md" to use a "Skill" (Computer Experience)
that was in the "db/seeds.rb" file, in place of "Attention to detail",
which is not.

Also made the steps in testing the "Semester" toggle in the admin
interface more generic by not specifying a specific semester, since
they may change over time.

https://umd-dit.atlassian.net/browse/LIBITD-2593